aider: Add the compiled executable file "backup" to .gitignore.
# Aider chat conversation:

USER: add the file called "backup" as this is a compiled executable

ASSISTANT: To add the compiled executable file "backup" to the list of *read-write* files, you need to upload the file to the chat. Please add the file "backup" to the chat.

USER: add the file called "backup" as this is a compiled executable to .gitignore

ASSISTANT: Sure, to ignore the compiled executable file "backup" in the `.gitignore` file, we can add it to the list.

Here is the *SEARCH/REPLACE block*:

.gitignore
```gitignore
<<<<<<< SEARCH
# Ignore Go related files
*.go
# Ignore Go binaries
/bin/
=======
# Ignore Go related files
*.go
# Ignore Go binaries
/bin/
# Ignore compiled executable
backup
>>>>>>> REPLACE
```

This will ignore the compiled executable file named "backup".